Extract ServerThread test class for reuse
