+++ src/us/kbase/workspace/test/WorkspaceServerThread.java
@@ -1,0 +1,20 @@
+package us.kbase.workspace.test;
+
+import us.kbase.workspace.WorkspaceServer;
+
+public class WorkspaceServerThread extends Thread {
+	private WorkspaceServer server;
+	
+	public WorkspaceServerThread(final WorkspaceServer server) {
+		this.server = server;
+	}
+	
+	public void run() {
+		try {
+			server.startupServer();
+		} catch (Exception e) {
+			System.err.println("Can't start server:");
+			e.printStackTrace();
+		}
+	}
+}
